Converting authentication servers to use the event bus. (WIP)
