+++ src/main/java/Logging/Model/ConsoleLogger.java
@@ -59,10 +59,12 @@ public class ConsoleLogger {
     public void log(JsonObject data) {
         if (enabled) {
             setColor(data);
+
             String text = data.encode()
                     .replaceAll("(\":\")", "=")
                     .replaceAll("[{}\"]", "")
                     .replaceAll(",", " ");
+
             AnsiConsole.out.println(text);
         }
     }
