--- src/main/java/Protocols/Handler.java
@@ -1,18 +1,0 @@
-package Protocols;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-
-/**
- * @author Robin Duda
- */
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Handler {
-    String value();
-    Access access() default Access.AUTHORIZE;
-}
