+++ src/main/java/Protocols/HandlerProvider.java
@@ -8,10 +8,21 @@ import Protocols.Exception.HandlerMissingException;
 /**
  * @author Robin Duda
  */
-public abstract class HandlerProvider {
+public class HandlerProvider {
     protected Logger logger;
     protected Protocol protocol;
+    private String address;
 
+    protected HandlerProvider(Class clazz, Logger logger, String address) {
+        this.protocol = new Protocol(clazz);
+        this.logger = logger;
+        this.address = address;
+    }
+
+    /**
+     * Processes an incoming request with authentication control.
+     * @param request the request to be processed.
+     */
     public void process(Request request) {
         try {
             protocol.handle(this, request);
@@ -22,5 +33,13 @@ public abstract class HandlerProvider {
 
             logger.onHandlerMissing(request.action());
         }
+    }
+
+    /**
+     * Get the address of which the handler is providing handlers for.
+     * @return the address as a string representation.
+     */
+    public String getAdddress() {
+        return address;
     }
 }
