+++ src/main/java/Protocols/Handles.java
@@ -1,0 +1,18 @@
+package Protocols;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+
+/**
+ * @author Robin Duda
+ */
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Handles {
+    String value();
+    Access access() default Access.AUTHORIZE;
+}
