--- src/main/java/Protocols/Message.java
@@ -1,7 +1,0 @@
-package Protocols;
-
-/**
- * @author Robin Duda
- */
-public class Message {
-}
