--- src/main/resources/1-vertx-default-jul-logging.properties
@@ -1,1 +1,0 @@
-handlers=Logging.Model.DefaultLogger
