+++ src/main/resources/vertx-default-jul-logging.properties
@@ -1,0 +1,1 @@
+handlers=Logging.Model.DefaultLogger
